Require that stderr is interactive when prompting
Because prompts to the user are sent on stderr (to allow pipelines of
commands), we need to require that stderr is a tty in order to get data out of
the user.
This restricts us from the 'echo ... | globuscli ...' use-case, but I think
that's okay for now. That should be entirely doable without accepting data from
stdin, which gets messy.