Add an `--additional-params` argument to all cmds
Also adds a `--supports-additional-params` argument to go alongside this, and a
decorator for setting that value to True on functions.
Basically, the behavior works by supplanting what will run during the
preprocessing of arguments (i.e. the stage between doing a raw argparse
parsing pass, and actually invoking the function described by argparse).
Instead of pulling out the func to run, just does an explicit check then and
there and exits.