Add initial version of auth subcommand
This is still something of a work in progress because token introspect
fails without handling for client credentials in a basic auth header.
For now, just raise a NotImplemented error, 'cause what else can we do?
All of the command parsing gymnastics are starting to feel a little bit
cumbersome -- starting to consider trying to add a bit more
sophistication; maybe a decorator or other bit of moderately magical
configuration.