Rename package and command per feedback
After some discussion, I'm renaming the package to `globus_cli` for consistency
with the `globus-sdk` way of translating dashes in the repo name to
underscores in the package name.
As an additional bit of copycating, changing from `globuscli` to `globus` as
the command, which is much like the `awscli` package providing the `aws`
command.