Change the way that commands are defined
Rather than using the Argparse constructs to define all of our numerous
arguments, which is rather tedious, put them all in a simple python dict
and go from there. As a dict (rather than JSON) it can have functions
rather than just their identifiers, tuples, and other objects. The
ultimate goal is to define only *exactly* the argument names, help
strings, and other textual elements that are needed, and not repeatedly
create subparsers with the same or nearly identical configuration,
replicated across services and commands. The goal is to remove two
levels of indirection from the process of mapping a value for `sys.argv`
to a function to run.