Add a --batch argument to async-transfer
Also adds the necessary logic for doing more complex/intensive
validation of arguments than what ArgParse provides. Adds an optional
argument to the @cliargs() decorator to hand it an argument validator
callable, which it will evaluate on any arguments. The advantage here is
that the application of the validator and its logic are kept separate
from the actual implementation of the desired command. Invocation is
also nicely uniform, taking place in the global command loading/parsing
code, before execution begins (so it can't be forgotten, if specified).