Add automagic autoactivate to Transfer operations
File ops and tasks trigger auto activation if the activated credential
expires in under 60s. For most cases, this is sufficient (e.x. GCP), and
it will be rare that this fails on an endpoint that *does* have an
active credential with a lifetime over one minute -- i.e. those cases in
which the task or operation would have worked fine.