Convert whole CLI to use click over argparse
It would have been nice to be able to do this in stages as a series of
commits, but it just didn't make sense -- there are very few functional
intermediate states for the CLI during this conversion process. All of
the sophistication in globus_cli.parser was necessary because ArgParse
doesn't actually have very good support for subcommands. It works when
the parsing and code execution are entirely distinct ideas, but it
starts to break down as soon as you want to tie the structure of the CLI
to the modular structure of the code itself. The decorators and classes
that I used were okay stopgaps, but not satisfactory long-term. The
problem was a lower-level model mismatch.
After careful evaluation, I chose click because of its superior support
for subcommand structures, its support for a wide range of python
versions, and its dedicated developer base (so it won't be abandonware in
a few short months). It also handles a number of py2 vs. py3
compatibility issues for us out of the box, which should save time and
heartache.
The click-based CLI has altered help strings, and required that we drop
the list-commands command (at least for now, investigate reviving it).