Promote helpers from module to package
The time has finally come -- helpers are too big and sprawling, and we
need to be able to find things other than the massive printing helpers.
To do this, make a bunch of small modules in a package and preserve
compatibility by exporting the same names.
The one exception is the CaseInsensitiveChoice type, which has been
moved into globus_cli.helpers.
Also restores flake8 passing status.