Add --recursive flag to `transfer ls`
Break up the `ls` function into a set of helpers, which is a kind of
decomposition necessary for a recursive and non-recursive ls to be
printed and handled in much the same way. Currently prints text-form
output in a format similar to `find`, rather than `ls -R` for
simplicity.
Tries to imitate the API behavior regarding slash termination, and
generally seems to plug in okay.