First version of endpoint server commands
Started with the easy ones -- list, show, and delete. Adding "add" and
"update" comes next.
Following the pattern of `endpoint role`, this is a set of subcommands
hanging off of `endpoint`, so we do `globus transfer endpoint server
list` and provide an endpoint ID.
For now, these are very simple, and don't even support non-json output
yet.