First pass at config commands
Initial effort on #12

Adds the `globus config` command family, with (so far) `set`, `show`,
and `edit`.
Introduces a dependency on configobj in order to load, edit, and write
config files without destroying comments.
Uses the dot notation familiar to us from `git config` to identify
section names, but supports omission of the section to default to the
"general" section.