Improve config inspection; use full config parser
Rather than using configobj (which is nice for editing single files),
inspecting values via `globus config show` should use the config parser
that is used for the SDK (a stdlib SafeConfigParser).
This also folds together well with having commands start actually
parsing and using values that are set for the CLI specifically -- namely
the output format for now.
To make this work, required a bit of renaming. Things in
globus_cli.config are now moved to globus_cli.config_command to make
room for the new globus_cli.config module, which reads the config files
in an SDK-like way.