Add batchmode to transfer async-delete
Batch input to async-delete is processed the same exact way that input
to async-transfer is, but with a different set of arguments and options.
In the case of deletes, it's actually just a single argument and no
options, but it's nice to use the same structure for both.
Adds a transfer helper method that generalizes that line-parsing
structure for use in both methods.

Resolves #4