Allow Task submission to take a submission ID
This adds a new command and a new option.
$ globus transfer task generate-submission-id
will create a new submission ID and print the result either raw or as
JSON.
As a result, you can now run
$ globus transfer async-transfer --submission-id ...
$ globus transfer async-delete --submission-id ...
to submit retryable Transfer Tasks.
Becasue the submission ID is generated in a separate step, it would be
nice to clean up the support for this in the SDK.

Resolves #14