Fix `globus config init` for new `globus login`
With the new form of the login command, there's no real point in having
the config init command consume access tokens -- that seems like an
antipattern that we should try to avoid in general, since the ATs aren't
a good credential to put in config.
It's not immediately clear what, other than setting your output format,
you might want to set manually in the config. For now it seems fine to
have `config init` do nothing else, but we may want to consider
expanding it later.