Improve arg parsing exception handling
The custom excepthook logic, if triggered on a Click exception, was
swallowing those exceptions. In fact, what we should do is check if the
exception is a descendant type of ClickException and re-raise if we're
not in debugmode. That way, we get the full stacktrace on Click
exceptions when we are in debugmode, and have the nice automatic
formatting of usage errors when not in debugmode.