Allow error printing to respect JSON format
When we are handling an error and running outside of debugmode, a number of
types of errors (e.x. usage errors) will result in raw text output. However,
for the SDK generated errors, we should now always be generating JSON data on
stderr if `--format JSON` or equivalent is passed.
Nicely, this means that stderr becomes parseable in certain applications (where
the caller is using the CLI correctly), and gives the formatting option a more
consistent meaning.