Expand JSON error handling to cover non-API errors
To capture globus_sdk.exc.GlobusError in the general (non-API) case, we
need to define another excepthook. The JSON form of the error will have
it stringified in one field, and the error class name in another.
Mostly just a matter of keeping the promise of JSON errors where
reasonable.