Reformat JSON error out to be indented
On the argument that it "looks better" and is more human-readable. In
the event that an error is produced, the user is not necessarily
expecting it (duh!). In that case, the assumption that they want to
feed the error output into parsing code could be wrong -- they may want
to inspect the output manually without necessarily specifying a
different output format (changing the command).