Update to handle multi-environment config
This is mostly an internal need, in order to run the CLI against Globus
development stacks, but it will also serve for the future beta
environment. Basically hooks into the SDK's config loader a little bit
more and uses the `GLOBUS_SDK_ENVIRONMENT` environment variable to
handle (1) loading the correct environment-specific Client ID and (2)
storing and retrieving refresh tokens with the environment as a prefix.

Also, bump version for development