Set default umask on config file to 077
The default umask for the config file when opened for config editing is
now restricted to user-only permissions. This applies both to `globus
config edit` and to `globus config init`, as well as other commands that
may write to the config file, like `globus login`.
Also, removes some of the duplication between globus_cli.config and
globus_cli.config_command
Resolves #25