Move all parsing code into globus_cli.parsing
Some option collections and other bits of special stuff that it seemed
to make sense to colocate with commands was being stored directly in
with the service code. However, the result is that we blur the line in
terms of how much sophisticated parsing code can live in which the
globus_cli.services subpackage.
To better differentiate, and keep the service-oriented bits well
focused, move all of this stuff into the parsing package. This may be
too extreme -- specialized options like `--role-id` only really make
sense in their relevant components (`globus transfer endpoint role`),
but the addition of the `share` commands demonstrated, fairly clearly,
that the same options or even collection of options may be used in
multiple places. Without this rigor, I caught myself thinking about
re-implementing the same or similar options in multiple places, which
is very bad.