Add macOS install to docs, move docs into tree
`master` will now track docs (we're going to try the new GitHub pages
style of HTML docs), so these are moving into this branch.
Also, adds some documentation on the macOS setup process, as the system
python on that platform is more ruinous (and the users are, frankly,
more likely to be naive).