Map exit codes by HTTP Status Mapping
This is the implementaiton half of the --map-http-status argument. For
each type of GlobusAPIError, we'll have to invoke the
"exit_with_mapped_status" helper on the http status.
Additionally, fix a bug in which other exception types (which should
always exit(1) ) were exiting with status 0 because we caught and did
not further handle exceptions. This bug was probably introduced at some
point around 0.3.x

There's a big question about how well the semantics of this will play
with auxiliary calls to the services like identity mapping or
autoactivation. Ideally, this would be hooked in more directly on the
error or response objects created by the "main" calls of the various CLI
commands.

Resolves #21