Let `transfer ls` default to default_directory
Simply remove the `path` parameter unless we have a value for it. Since
it's not obvious that the Transfer API will always treat `path: null`
equivalently to the path being absent, use kwarg-selection to only
sometimes pass it to the SDK.
(I think the SDK handles path=None by removing it too, but this way is
safest).
Fixes #49