Make async-* ops do colon format, show task_id
async-delete and async-transfer were doing raw "message" output (which I
think we want to get away from in general? unsure). Instead, use
colon_formatted_print to do the output handling, and use it to show the
Task ID for the submitted tasks. Much better for submitting a task and
then interacting with it.
Fixes #29