Add EndpointPlusPath parameter type, convert ls
`globus transfer ls` is the first command to get the new treatment,
taking an EndpointPlusPath positional parameter. This is a custom type
which parses "ENDPOINT_ID[:PATH]"
When path is required, empty paths are the same as no path and fail.
When path is not required, empty paths are converted to NoneType.

The minor hack is that it supplies its metavar in an additional
property. I've filed a bug report against click to fix the underlying
behavior that makes this necessary.

Resolves #73