Bugfix: bookmark create with new UUID type
Was passing a UUID in the datadict, rather than stringifying it first.
Simple fix, just catch the value and `str()` it here.
Unlike the general acceptance of UUIDs by the TransferClient, I think
this makes sense to handle externally. Trying to handle it in the SDK
would mean doing inspection of a value that we've been passed, and
either carefully copying it (effortful) or mutating it (mildly
dangerous). For example, we can't very well rewrite the dict we're given
in the SDK from a UUID to a string, but the calling code knows whether
or not that's a safe thing to do.