Convert task_id_option to task_id_arg
Everywhere that takes `--task-id` now, instead, takes the `TASK_ID`
argument. For future suspected needs (like `task cancel --all`), takes
required=T/F as an argument to the decorator.

Also, break decorator application which can take args or not (i.e. has
parens and keyword args or has no parens) into a helper that lives
inside of parsing. The reason is simply that the family of functions
with this behavior keeps growing.