Update search to make filter_fulltext positional
This is a positional argument now, but still optional. Because the vast
majority of searches will use this value, it makes sense for it to be a
positional arg. However, by making it the only such argument, and
leaving it optional, we allow for searches that omit it and specify
other values to search (like scope and owner).