Add experimental tab completion
Supports both BASH and ZSH, implemented as a hidden option to the top
level `globus` command. The result is that as long as `globus` can be
invoked in your environment, so can `globus` tab completion -- a
desirable property for portability and safety.
This strategy is in place of the `_GLOBUS_COMPLETE=...` approach, which
doesn't appear to offer the same range of flexibility with the same
level of ease.

Also include a couple of scripts that can be used for completion
support in these shells. Sourcing them will have the desired effect.

This is a more desirable approach for us than trying to use the click
completion because (1) it's fully customizable, letting us dispatch on
types like the CaseInsensitiveChoice and doing "the right thing", (2)
click maintainers have called their existing support a "hack" and don't
seem very confident that it will always work, and (3) we can add
helptext for shells which support it in the future.
I also, personally, want to have it work in my shell (zsh), and have the
option of supporting other shells if their completion formats are simple
as well.