Join paths as consumed on task batchmode input
Per #89, slapping the commandline paths into the document body as just
another pair is potentially confusing and not very good semantically. To
be more useful, join paths as given on the commandline to all of the
batch inputs, letting the commandline paths be used to specify a prefix.
This leads logically towards an interesting use case for
`globus transfer ls | globus transfer async-delete ...`
That's also interesting for `async-transfer`, but requires the annoying
extra step of `| awk '{print $1 " " $1}'. The solution is simple -- let
`dest_path` be optional during transfer submission, and we'll just copy
the source path, potentially joined with a commandline dest path. With
this improvement, it's possible to pipe and `ls` through a grep,
potentially prepend `--recursive` on lines that end in `/`, and finally
pipe to transfer submission.

For example,
$ globus transfer ls --all "$GO_EP1" | grep '^.' | \
    globus transfer async_transfer "$GO_EP1" "$GO_EP2:ep1dotfiles/"

Although you can do
  awk '/\/$/{print "--recursive " $1} /[^\/]$/{print $1}' \
to handle dirs, at that point printing the filename twice is no big
deal.

This actually works (whoo!), which is pretty gosh-darn cool.
Importantly, I think this aligns well with a use case of users who will
be able to generate a list of absolute paths which they want to sync
across servers. Assuming those listings are all files or all dirs, we
can now consume them directly, which is actually desirable.

It also makes for better interactive use with globbing and grepping.
More viable filters to contend with `grep '^.'` are things like
`grep '\.txt$'` Supporting "transfer all of my TIFFs from A to B" is
much simpler with this.

Resolves #89