Convert security principal collection opts
Rather than having a pair of args, "--principal" and "--principal-type",
which are both required and hard to explain (especially since principal
isn't really required if the type is anonymous or authenticated users),
add a collection of flags to describe this behavior.
So, `--identity <ID>` or `--group <ID>` or `--anonymous` or
`--all-authenticated` as options, only some of which take args.
The callback functions and default values ensure that if none of them
are passed, we get None, but otherwise we get a tuple of
(principal_type, value) where value is None for those that don't take a
value.