Start rearranging for command tree refactor
Does the initial and major, noisy step of moving things into a
`commands` subpackage. We discussed that this wasn't necessary, but as
top-level modules are about to multiply significantly it feels like a
valuable first step.
Only two commands are actually changing locations in this commit:
`get-identities` and `ls` as proofs of concept. Everything else remains
as it was.