Update logout to better handle error conditions
Allows it to handle no username much better by prompting if it isn't
found. Additionally, fixes an error in remove_option() which causes it
to fail if a value is already absent/removed.
Adds a `--yes` option to logout to skip prompts.
Fixes opts on logout and also on login to not include irrelevant common
options.
Also change the order of operations to ensure we don't delete tokens
which we then can't revoke because of NetworkErrors.