Update logout command to print epilog text
Also warns if it sees missing tokens (nothing else is quite as
important, so that's all that it warns about), and prints a help message
about rescinding consents if any tokens were missing.
So far, missing tokens, and therefore failing revocation, is the only
viable case I can conceive of for wanting to rescind the consent.