PrintableErrorField: workaround self.value is None
Sometimes if an API provides a field but its value is `null`, this will
show up here as `self.value` being `None`, which breaks the check to see
if it contains a newline. We can avoid this by making sure that
`self.value` is indeed a string before making this check.