Fix `globus whoami` to allow JSON output
Simply put, JSON implies verbose, and the verbose document gets wrapped
in a GlobusResponse to be dumped to stdout.
Convert the "you are not logged in" message from being stdout + exit(0)
to behaving like an error message, going to stderr + exit(1). That way,
we have good behavior for `-F json` on that error.

Closes #206