Fold common_options into globus_group
This doesn't modify the underlying class, so if you wanted to customize
the application of common_options() in the future, it's still possible.
However, when a group is decorated with @globus_group(...), it was
always also being decorated with @common_options (bare, no args). Rather
than persist this, just do away with it by including one within the
other.