Improve handling of exceptions in local HTTP login
Two parts of this.

Part I: Make server start/shutdown a contextmanager.
This means that if an exception is raised, we're still guaranteed to
invoke shutdown().

Part II: If an error occurs inside of the HTTPServer handler, propagate
it out to the calling context.
Trying to invoke shutdown() inside of an error handler gets kind of
loopy because shutdown() waits for request processing to terminate.
Instead, abuse the auth code queue -- push any raised exception into
there, and check the result to see if it's an Exception before returning
(or reraising) it. Piggy back a bit on the usage already in place for
LocalServerError, but preserve the original exception and change the
message slightly to make it identifiably different.