Set encoding=utf-8 on ConfigObj
This is used for loading and writing ~/.globus.cfg
Setting the encoding allows ConfigObj to correctly handle unicode
strings and does not conflict with existing file contents in ascii-only
files.
Tested manually -- it's not clear at present how this should be added to
the testsuite, since tests currently run against _real_ configuration
data.