Cert validity timestamps now use UTC
If you live in GMT-plus-something land (Europe or Asia) the certs generated by create_proxy_cert() will not be valid for some hours because it inserts the local time. But GMT+N is for UTC "The Future". Having a "not yet valid" cert sucks really hard. Now it will use UTC to generate the cert and everything is rainbows.