Improve task event-details data in text output
When text output is requested from `task event-details`, we currently
dump the `details` field from the API as JSON. But critically, we do it
with indentation and spacing, including newlines, which make the tabular
output... not tabular anymore. This is actually a result of the fact
that `details` is a *string* output which is *typically* json.
Squash this output down to have no whitespace when it is JSON.
Otherwise, convert `\n` to `\\n` so that output will appear on one line,
but still encodes any newlines. (Corner case: if the event contained a
line ending in `\`, this would probably not render correctly, but it's
not really clear how good *any* of our text output is in the presence of
arbitrary control characters and `\` chars in the API data.)