Improve docs for `transfer --batch --recursive`
These options are mutually exclusive, but there's a `--recursive`
option inside of `transfer --batch`, which makes this a little confusing
if you don't realize that it's part of a different option set.
Try to clarify the docs around this a little, and also add a little more
meat to the examples section. Specifically, show a second `--recursive`
example, and an example which uses the `--batch` path prefixing
behavior.

Closes #370, #341