`globus delete` prompt on trailing globs
When used for an interactive command, `globus delete`, given an
endpoint-plus-path on the command-line, will prompt before submitting a
task with a trailing "*".
`--star-silent, --unsafe` can be used to suppress this behavior,
resulting in no prompt.

Interactivity is checked explicitly, ensuring that stderr, where the
prompt will be printed, is a pty. Add a small module to safeio for
handling the pty checks on stderr, stdout (though we only use one here).

Also, use PS1 env var to check if use is interactive.

Closes #129