Fix completion for bash with `--format=text`
Using COMP_WORDS and COMP_CWORD for completion in bash gives incorrect
results with the typical COMP_WORDBREAKS. Because `=` is considered a
word break, the completer will see `--format = text` instead of
`--format=text` and choke on it (that's not valid usage, after all!).
Instead of doing this, take COMP_LINE and use COMP_POINT. Things get a
little bit hairy because we need to shlex the line and grab the last
word as the current word. That lets us properly handle the current,
potentially incomplete, word.
Also, switch the completion options up a little bit -- set `nospace` to
ensure that `--format=<tab>` doesn't insert a space where it is not
wanted. This still has good behavior on normal options, and we can use
the presence of a space to indicate that the prior word was completed.