Add "secondary opts" to option completion
When multiple options are given to click in `--opt/--no-opt` notation,
the first option is passed into "Option.opts" and the second option is
passed into "Option.secondary_opts". The code we had was just iterating
over `opts`, but it needs to walk both.
Having the same description for both options has a beneficial effect in
ZSH completion, as the shell is smart enough to coalesce these, and show
both options with a single unified description (thus clarifying that
they are inverses controlling the same behavior in that context).