Replace completer scripts with `--completer` opts
`--completer/--bash-completer` and `--zsh-completer` are new hidden
options which print the completers for these shells.
They can be used, and are now shimmed in the old completion scripts,
via `eval "$(...)"` during shell initialization.
The completer funcs still check `type globus` to ensure that the CLI is
installed. If it's absent, we don't want to break the shell by
installing a badly broken completion function.