Install cryptography by default
Use cryptography>=1.8.1,<2.5.0 , to fit the deprecation policy
established in the cryptography packaged documentation.
cryptography does not use typical semantic versioning:
https://cryptography.io/en/latest/api-stability/

Importantly, this will now install wheel based distributions for most
platforms (leading to a faster, more reliable install).

Deprecate delegate-proxy extra install target, but leave it, empty.

Make all cryptography-conditional code non-conditional because
cryptography is now always-on.

Closes #351

NOTE:
Although the cryptography package promises strong compatibility, it also
sets a deprecation policy by which it could, in theory, drop features
which we are using and rely upon. In order to guard against this, we
set a reasonably conservative version constraint. We can update this
constraint and release new versions of the CLI over time, to improve our
compatibility with other `cryptography`-based projects.

Let us not lie to ourselves: the setup.py will rarely, if ever, be
updated for this purpose, especially as we would generally recommend
isolating the CLI into its own virtualenv. However, we may want to
include newer dependencies ourselves, or improve support for `--user` or
system installs.