Fix `globus update` under `pip install --user`
Using `site.USER_BASE` and checking on `__file__`, we can reasonably
reliably detect whether or not a `--user` install was done when the CLI
was originally installed. (There are always bizarre situations, custom
userbase, etc. But this works for the "normal" case and advanced users
are advanced.)

Once we do that, it's a very simple matter to change
  `pip install --upgrade globus-cli`
to
  `pip install --user --upgrade globus-cli`

Also, print the pip command now, since it will help us troubleshoot when
people have issues.