Fix invocation of `pip` to follow guidance
The `pip` documentation explicitly states that calling `pip.main()` is
not allowed, and that calling via `subprocess` + `sys.executable` *is*
the officially supported mode of use for invoking `pip` within a running
python process.

We still `import pip` and assert that `pip` can be found so that we can
error out early and provide good feedback when pip is missing.
Otherwise, we'd get to the point of execution and error -- at that
point, it's a bit too late to provide a good error message.