Add `globus rm` and refactor the supporting cast
`globus rm` features all of the behaviors of `globus delete` except for
`--batch` mode, and *all* of the behaviors of `globus task wait`.
Especially the most important one. If you've seen the source, you know
which one.

Like `globus task wait`, we'll happily abuse stderr for all of our
cosmetic IO and let stdout only be used when `-Fjson` is asked for. (In
other words, the pretty text output is always "").

Refactor of support went as follows:
- Make options for `globus delete` and `globus task wait` into "shared
  option" decorators in globus_cli.parsing
- Move the main loop and all IO capabilities of `globus task wait` into
  globus_cli.services.transfer , including the exiting behaviors
- Allow for a big DRY violation between `globus delete` and `globus rm`
  -- is there a clean and obvious way of resolving this?

Closes #196