Add support for `--format=unix`
Pull in the awscli `--output=text` formatter, modify it very slightly to
make it usable, clearly marking modifications, and setup `--format=unix`
to pass `--jmespath` processed data through this transform.

`--format=unix` is not guaranteed to be pretty and friendly, and with
certain nesting structures can actually be quite a mess. But it is
*consistent* and *good enough*.

The data is output as TSV, and we're okay with that. Fields containing
tabs may expand in nasty ways -- but that's pretty rare and people
shouldn't do that anyway. If you need really careful data handling, use
the SDK or the `-Fjson` output anyway.