Add pretty error when --format=unix dies
The "UNIX" formatted output can fail when you give it certain datatypes,
and we don't want to try to do too much. Simply use a try-except to
catch AttributeErrors.

Because the formatter does writes to stdout as it works, there may be
partial data on stdout. That's *fine*.
The alternative is fully-buffered output, which is slower and offers no
clear benefit other than being "prettier" in obscure error scenarios.