Make JSON output use `sort_keys` to force ordering
When dumping an object to JSON output, use `sort_keys` to ensure that
ordering is consistent.

This doesn't need to be a documented part of the interface for the CLI,
that we use sorted JSON output, but it results in more consistent
behavior. Nobody *should* be relying on the order of the output, but
this way, if they *are*, we'll be able to give them better behavior.