`get-identities` autodetects b32-encoded users
Get rid of `_HIDDEN_TRANSFER_STYLE` and the `--globus-transfer-decode`
flag altogether. Instead, have `get-identities` inspect the strings it
receives and try to decode. This makes the failure modes for malformed
b32 names a little dirtier, but it makes it a lot easier to explain to
users how to handle the b32 names without having to get into a history
of the Auth rollout and why Transfer needs unix usernames.

This seems increasingly valuable as it turns out that GCS logging can
encode the Sharee ID in this format, showing it directly to endpoint
admins.