Batch resolve IDs to usernames in permissions list
When doing an ACL listing, we resolve identities one-by-one (*and* we
don't keep the results). The result is that the permissions list takes
many times longer than it should.

For now, apply a fix directly in this command. As a future refinement,
we should make mapping identities to usernames in bulk a generic
operation which many commands can share. There are presumably other
cases for this type of lookup? If not, leave it as something specific to
this command, but maybe come back and clean it up a little.

I'd like to do something in which we can be a bit less ham-fisted about
checking for text output, but maybe this way is best.

Closes #422