Simplify lazy bulk identity lookups
Define a new helper object: LazyIdentityMap which is dict-like (but
not actually a dict -- why bother?) which handles mapping identity ids
to names.

It accepts a list of IDs to lookup when it's instantiated, and does the
lookup whenever it is first accessed. This isn't the "laziest" behavior
possible, certainly -- but it's simple and effective.

The big advantage is that we can declare that we want the
LazyIdentityMap ready, and then not worry about which output format is
going to be used. It will only (in practice) be checked when text
formatting is done.

As a result, `-Fjson` and `-Funix` output still run just as fast, and
`-Ftext` output triggers the right action, but without a sloppy
`outformat_is_text()` check in the command to decide whether or not to
populate an identity map.