Update tests to pass on Windows, py2 and py3
Numerous small tweaks to the tests to pass on Windows. Includes some
hacks which we should remove when #455 is fixed, all of which are
carefully documented as such.

The current work in progress is not about compatibility for the
testsuite, so rather than trying to fix Windows python2 console issues,
these are all marked as skip for now.