Replace safeprint() with click.echo()
Now that Click handles EPIPE internally, we don't need to wrap all
click.echo() usage.

Note that click.echo does not do the EPIPE handling, but rather the
Click context teardown catches EPIPE which might arise from any
click.echo() in the CLI. So this will not be visible behavior if you try
to test click.echo directly -- you need to be running a full click
application to see it.