Restore blanket exc handling, catch click Exit
Restore the behavior that an uncaught exception gets "pretty-printed",
but extend the search for click exceptions to cover
click.exceptions.Abort and click.exceptions.Exit

These internal exceptions are meant to signal exit (e.g. ctx.exit(...)
calls) to the main click CLI logic. We need to match on them explicitly
so that they don't land in our catchall handler.

The catch-all handler shows up in contexts like UnicodeDecodeError on
invalid inputs, so it's actually quite important to providing a good
UX.