Revert jmespath ParseError behavior
In one of the intermediate commits of #465, JMESPath ParseError handling
had to be reworked a bit to handle the top-level exception handling
being changed. But this change actually results in more verbose, worse
errors for the most part. It's better to handle these the way we were
doing before, so this reverts the change.