Fix path normalization on Windows
closes #455

batchmode transfers and deletes from a Windows system should still
produce POSX-style paths. Otherwise, these will join incorrectly on
Windows and then be passed to Transfer where they will be treated as
POSIX paths. In order to handle this nicely without much headache,
simply define our own normpath and pathjoin functions based on the
posixpath implementation in the python stdlib.

This should *NOT* be thought of as "POSIX normalization", but rather
"Globus Transfer normalization". While in practice those two may be the
same, there are in theory subtle differences.

For example, this implementation takes the liberty of not caring about
the number of leading slashes on a path, on the assumption that this can
hardly even (if ever) matter to Globus Transfer.

Testsuite changes:

Now that #455 is being fixed, the specific tweaking we did to make tests
pass on Windows should be undone.

Using the dry-run delete tests, add another case with some weirder paths
(leading slash, `/./`, and `..` use), and make sure that when joined
with the given base path and normalized, the expected values show up.