Flatten out the testsuite into fewer directories
We don't have a meaningful distinction between "integration" tests and
"unit" tests in our current testsuite. The result is that we have things
somewhat randomly fragmented across directories.
Additionally, the `tests.framework` subpackage could just be promoted
up.

A bunch of imports needed rewriting, but the result should be a bit
simpler to read and understand.