Begin splitting testsuite into unit vs functional
Functional tests talk to the live API. Unit tests are small and fast,
safe to run many times repeatedly. They're how we can drive up our
test coverage without having to extensively mock over the API surface.

Mocking an API surface and testing against it would also qualify in
this breakdown as a "unit" test. The idea is to split the live,
functional testing from anything that's purely local, in-memory,
testing. There are some things in the "functional" part of the suite
which could individually move over, but these aren't being pursued at
the moment.

For now, split and add a single test module for testing a bookmark
handling helper. (Just to prove the idea.)