Update Travis to not run linting on each platform
Windows linting failures are a dumb way for our builds to be marked as
failed. Just run linting on the linux worker, TYVM.

It's conceivable for the linting to fail on Windows or OSX due to
strangeness -- in fact, that just happened -- but this isn't worth
digging into for us. So long as things aren't badly broken, we can
consider variation by platform in the "lint" target to be a bug or
issue with the linters/fixers.