Pull command helptext from docstrings
Add a pair of custom decorators, `command` and `group`, which mirror
`click.command` and `click.group` but pull `help` from the docstring
of the function if `help` isn't supplied explicitly. (So you still get
a choice to do things the other way if you like.)
`group` is just a change to the already existing `globus_group`.

Move all commands and groups to using these decorators, and all
helptext into docstrings, eliminating the pointless and redundant
"executor for ..." docstrings we padded in long ago.

Also rewrites handling of common_options in two major ways:

1. Supplying common_options is no longer required -- it's part of
   `command`.
2. Disabling options is done by passing a list, not super-long
   keyword arguments.

`common_options` is no longer exposed directly by the parsing package,
so it's very well-contained now.

Net-net, this reduces the size of our codebase and amount of
redundancy with no significant downsides.