Store refresh tokens whenever a refresh happens
It is possible for a new refresh token to be sent back upon refresh.
Per the OAuth spec (RFC 6749):

  The authorization server MAY issue a new refresh token, in which case
  the client MUST discard the old refresh token and replace it with the
  new refresh token.
                        https://tools.ietf.org/html/rfc6749#section-6

So we are technically in violation of the spec right now. We happen to
know that Auth does not currently rotate refresh tokens, so it isn't
an issue, but we'd like to be a spec-compliant client so that there
aren't issues if this part of the spec is exercised in the future.

It is reasonable to believe that refresh token expiration may become
mandatory in the future -- in preparation for the HA featureset, the
existence of long-lived refresh tokens was noted as a gap in our
credential management processes.

Implementation-wise, the change is quite simple: the callback helpers
which write new access tokens to config now also require the refresh
token and rewrite it each time. Some mild name changes to be accurate,
and it's done.
Manual tests confirm that we still write correct results from the
handler.

fixes #414